+++ common/src/main/java/dev/corgitaco/worldviewer/client/tile/tilelayer/NoiseCaveLayer.java
@@ -97,9 +97,9 @@ public class NoiseCaveLayer extends TileLayer {
 
                 if (foundCaveBlocks > 2) {
                     int grayScale = getGrayScale(((float) foundCaveBlocks) / ((float) searchRange), dataTileManager.serverLevel());
-                    image[sampleX + sampleZ * sampledSize] = (byte) grayScale;
+                    image[sampleX + sampleZ * sampledSize] = grayScale;
                 } else {
-                    image[sampleX + sampleZ * sampledSize] = (byte) FastColor.ABGR32.color(0, 0, 0, 0);
+                    image[sampleX + sampleZ * sampledSize] =  FastColor.ABGR32.color(0, 0, 0, 0);
 
                 }
             }
